Dev (#1)
Cleanup, Go over TODOs, finalize tests

+++ recipes-kotlin-spring-ddd/src/main/kotlin/io/immortal/spicemustflow/domain/recipe/RecipeIngredient.kt
@@ -1,8 +1,9 @@
 package io.immortal.spicemustflow.domain.recipe
 
 import io.immortal.spicemustflow.domain.ingredient.IngredientId
+import java.math.BigDecimal
 
 data class RecipeIngredient(
     val ingredientId: IngredientId,
-    val amount: Double
+    val amount: BigDecimal
 )
